+++ types/gen/go/transport/v1/transport.pb.validate.go
@@ -2075,6 +2075,8 @@ func (m *Request_Data_Data_Props) validate(all bool) error {
 
 	// no validation rules for Version
 
+	// no validation rules for UseWasmBindingsSandbox
+
 	if len(errors) > 0 {
 		return Request_Data_Data_PropsMultiError(errors)
 	}