+++ workers/javascript/packages/server/src/language-step-executor/jsLanguageStepExecutor.ts
@@ -138,6 +138,8 @@ export class JsLanguageStepExecutor implements Closer {
     pluginProps.context.variables = pluginProps.variables ?? {};
     pluginProps.redactedContext.kvStore = this._kvStore;
     pluginProps.redactedContext.variables = pluginProps.variables ?? {};
+    pluginProps.context.useWasmBindingsSandbox = pluginProps.useWasmBindingsSandbox;
+    pluginProps.redactedContext.useWasmBindingsSandbox = pluginProps.useWasmBindingsSandbox;
     pluginProps.version = request.props.version;
 
     pluginProps.actionConfiguration = this._convertActionConfig(pluginProps.actionConfiguration, observabilityTags);